+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - **BREAKING:** Rename `RumbleInterface` to `RumbleDriver`.
+- **BREAKING:** Rename `GeneratorInputDriver` to `IterableInputDriver`.
 
 ## [0.3.1] - 2024-10-17